backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
spacy==3.7.2
transformers==4.35.2
torch==2.1.1
nltk==3.8.1
PyMuPDF==1.23.8
python-docx==1.1.0
python-dotenv==1.0.0
pydantic==2.5.0
sentencepiece==0.1.99